Repair error handling with requests.
This was originally written to use urllib2, which would raise HTTPError for
non-200 response. This is a special exception which can be read like an
open file to obtain the body of the response. That's why AdsAPIError would
call json.load(error).

To handle this properly with requests, it's easier to decode in the caller
and instantiate AdsAPIError with the data.